Added confirmation functionality when user - deletes CV, delete announced job, announce a job. User cannot choose date less then current when announcing new job.

+++ src/com/clouway/jobex/client/cv/UserCVsPresenter.java
@@ -105,6 +105,11 @@ public class UserCVsPresenter extends AbstractActivity implements ApplyForJobEve
   }
 
   public void deleteCv(long cvId) {
+
+    if (!view.isConfirmed()) {
+      return;
+    }
+
     JobExRequestFactory.CVsRequestContext context = requestFactory.cvsRequestContext();
 
     context.delete(provider.getUsername(), cvId).fire(new Receiver<List<CVProxy>>() {
